refactor: extract shared pagination utilities to reduce duplication
- Extract parseNextStart() to src/utils/pagination.ts (was duplicated 4x)
- Extract parsePaginationOptions() for limit/start validation (was duplicated 2x)
- Replace inline copies in attachments, comments, properties clients + commands
- All 83 tests pass, lint clean, typecheck clean
